backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
django-extra-fields==3.0.2
psycopg2-binary==2.9.7
reportlab==4.0.4
python-dotenv==0.21.1
django-filter==23.2